+++ testsuite/benches/w3c_testsuite.rs
@@ -84,7 +84,7 @@ fn parse_bench(
 
 fn parse_ntriples(c: &mut Criterion, name: &str, data: Vec<u8>) {
     parse_bench(c, "ntriples", name, data, |data| {
-        let mut count: usize = 0;
+        let mut count: u64 = 0;
         NTriplesParser::new(data)
             .parse_all(&mut |_| {
                 count += 1;
@@ -96,7 +96,7 @@ fn parse_ntriples(c: &mut Criterion, name: &str, data: Vec<u8>) {
 
 fn parse_nquads(c: &mut Criterion, name: &str, data: Vec<u8>) {
     parse_bench(c, "nquads", name, data, |data| {
-        let mut count: usize = 0;
+        let mut count: u64 = 0;
         NQuadsParser::new(data)
             .parse_all(&mut |_| {
                 count += 1;
@@ -109,7 +109,7 @@ fn parse_nquads(c: &mut Criterion, name: &str, data: Vec<u8>) {
 fn parse_turtle(c: &mut Criterion, name: &str, data: Vec<u8>) {
     let base_iri = Iri::parse("http://example.com/ex".to_owned()).unwrap();
     parse_bench(c, "turtle", name, data, |data| {
-        let mut count: usize = 0;
+        let mut count: u64 = 0;
         TurtleParser::new(data, Some(base_iri.clone()))
             .parse_all(&mut |_| {
                 count += 1;
@@ -122,7 +122,7 @@ fn parse_turtle(c: &mut Criterion, name: &str, data: Vec<u8>) {
 fn parse_trig(c: &mut Criterion, name: &str, data: Vec<u8>) {
     let base_iri = Iri::parse("http://example.com/ex".to_owned()).unwrap();
     parse_bench(c, "trig", name, data, |data| {
-        let mut count: usize = 0;
+        let mut count: u64 = 0;
         TriGParser::new(data, Some(base_iri.clone()))
             .parse_all(&mut |_| {
                 count += 1;
@@ -135,7 +135,7 @@ fn parse_trig(c: &mut Criterion, name: &str, data: Vec<u8>) {
 #[cfg(feature = "generalized")]
 fn parse_gtrig(c: &mut Criterion, name: &str, data: Vec<u8>) {
     parse_bench(c, "gtrig", name, data, |data| {
-        let mut count: usize = 0;
+        let mut count: u64 = 0;
         GTriGParser::new(
             data,
             Some(Iri::parse("http://example.org/base/".to_owned()).unwrap()),
